test: move url_upload tests to fix coverage
Moved `test_url_upload_redirect.py` into `tests/test_url_upload.py` to ensure it is picked up by the coverage suite.
